Added the farmer's dashboard page
