+++ public/css/style.css
@@ -118,7 +118,7 @@ h6 {
 .nav-menu a {
   display: block;
   position: relative;
-  color: black;
+  color: pink;
   transition: 0.3s;
   font-size: 15px;
   padding: 0 4px;
