+++ src/App.js
@@ -6,6 +6,7 @@ import { BrowserRouter as Router, Route,Switch } from "react-router-dom";
 import Home from "./Components/Pages/Home";
 import FarmerDashboard from './Components/Pages/FarmerDashboard';
 import EditFarmerProfile from './Components/Pages/EditFarmerProfile';
+import FarmProduce from './Components/Pages/FarmProduce';
 
 const App = () => {
   return (
@@ -15,6 +16,7 @@ const App = () => {
           <Route exact={true} path="/" component={Home} />
           <Route exact={true} path="/dashboard" component={FarmerDashboard} />
           <Route  path="/dashboard/edit" component={EditFarmerProfile} />
+          <Route  path="/dashboard/products" component={FarmProduce} />
         </Switch>
     </Router>
   );
