+++ src/styles/SideNav.css
@@ -1,0 +1,22 @@
+ .navigation {
+    height: 100%;
+    background: black;
+    color: white;
+    margin: 1rem auto 0 auto;
+    padding: 4%;
+} 
+li {
+    list-style: none;
+} 
+hr {
+    border: 1px solid pink;
+}
+
+.navigation ul {
+    padding-inline-start: 0;
+    margin-bottom: 0;
+} 
+.img-fluid {
+    border: 2px solid white;
+    padding:2px;
+}
